+++ lsm-tree/src-core/Database/LSMTree/Internal/Unsafe.hs
@@ -77,6 +77,7 @@ module Database.LSMTree.Internal.Unsafe (
   , readCursorWhile
     -- * Snapshots
   , SnapshotLabel
+  , SnapshotMode (..)
   , saveSnapshot
   , openTableFromSnapshot
   , deleteSnapshot
@@ -1913,26 +1914,70 @@ listSnapshots sesh = do
       if b then pure $ Just snap
             else pure $ Nothing
 
+{- |
+The mode to be used by a snapshot import or export.
+-}
+data SnapshotMode m h
+  = HardLink
+    -- | Whether or not to allow fallback to copying.
+    !Bool
+  | Copy
+    -- | The 'HasFS' instance used for copying.
+    !(HasFS m h)
+
+{- |
+Internal helper.
+
+Get an 'FsErrorPath' for an 'FsPath' interpreted under the given 'SnapshotMode'.
+-}
+snapshotFsErrorPath ::
+     HasFS m h
+  -> SnapshotMode m h'
+  -> FsPath
+  -> FsErrorPath
+snapshotFsErrorPath hfs mode fsPath =
+  case mode of
+    HardLink {} -> FS.mkFsErrorPath hfs fsPath
+    Copy hfs'   -> FS.mkFsErrorPath hfs' fsPath
+
+{- |
+Internal helper.
+
+Convert a 'SnapshotMode' to the internal 'FS.Mode'.
+-}
+fsMode ::
+     HasBlockIO m h
+  -> SnapshotMode m h'
+  -> FS.Mode m h
+fsMode hbio = \case
+  HardLink fallback ->
+    FS.HardLink fallback hbio
+  Copy hfs ->
+    FS.Copy hfs
+
 -- | A snapshot was intended to be imported, but the source directory does not
 -- exist.
-data SnapshotImportDirDoesNotExistError
-    = SnapshotImportDirDoesNotExistError !FsPath
+newtype SnapshotImportDirDoesNotExistError
+    = ErrSnapshotImportDirDoesNotExist FsErrorPath
     deriving stock (Show, Eq)
     deriving anyclass (Exception)
 
 {-# SPECIALISE importSnapshot ::
      Session IO h
   -> SnapshotName
+  -> SnapshotMode IO h'
   -> FsPath
   -> IO () #-}
 -- |  See 'Database.LSMTree.importSnapshot'.
 importSnapshot ::
+  forall m h h'.
      (MonadMask m, MonadSTM m, PrimMonad m)
   => Session m h
   -> SnapshotName
+  -> SnapshotMode m h'
   -> FsPath
   -> m ()
-importSnapshot sesh snap sourcePath = do
+importSnapshot sesh snap mode sourcePath = do
     traceWith sesh.sessionTracer $ TraceImportSnapshot snap sourcePath
     withKeepSessionOpen sesh $ \seshEnv ->
       withActionRegistry $ \reg -> do
@@ -1947,7 +1992,7 @@ importSnapshot sesh snap sourcePath = do
         let destinationPath = Paths.getNamedSnapshotDir snapDir
 
         sourceExists <- FS.doesDirectoryExist hfs sourcePath
-        unless sourceExists $ throwIO (SnapshotImportDirDoesNotExistError sourcePath)
+        unless sourceExists $ throwIO $ ErrSnapshotImportDirDoesNotExist $ snapshotFsErrorPath hfs mode sourcePath
 
         -- we assume the snapshots directory already exists, so we just have
         -- to create the directory for this specific snapshot.
@@ -1955,8 +2000,8 @@ importSnapshot sesh snap sourcePath = do
           (FS.createDirectory hfs destinationPath)
           (FS.removeDirectoryRecursive hfs destinationPath)
 
-        -- create hard links for all files in the destination directory
-        FS.hardLinkDirectoryRecursive hfs hbio reg sourcePath destinationPath
+        -- import the files for the snapshot, either by hard linking or copying
+        FS.hardLinkOrCopyDirectoryRecursive hfs (fsMode hbio mode) reg sourcePath destinationPath
 
         -- Make the destination directory and its contents durable
         FS.synchroniseDirectoryRecursive hfs hbio destinationPath
@@ -1968,14 +2013,15 @@ importSnapshot sesh snap sourcePath = do
 
 -- | A snapshot was intended to be exported, but the destination directory
 -- already exists.
-data SnapshotExportDirExistsError
-    = SnapshotExportDirExistsError !FsPath
+newtype SnapshotExportDirExistsError
+    = ErrSnapshotExportDirExists FsErrorPath
     deriving stock (Show, Eq)
     deriving anyclass (Exception)
 
 {-# SPECIALISE exportSnapshot ::
      Session IO h
   -> SnapshotName
+  -> SnapshotMode IO h'
   -> FsPath
   -> IO () #-}
 -- |  See 'Database.LSMTree.exportSnapshot'.
@@ -1983,9 +2029,10 @@ exportSnapshot ::
      (MonadMask m, MonadSTM m, PrimMonad m)
   => Session m h
   -> SnapshotName
+  -> SnapshotMode m h'
   -> FsPath
   -> m ()
-exportSnapshot sesh snap destinationPath = do
+exportSnapshot sesh snap mode destinationPath = do
     traceWith (sessionTracer sesh) $ TraceExportSnapshot snap destinationPath
     withKeepSessionOpen sesh $ \seshEnv ->
       withActionRegistry $ \reg -> do
@@ -2000,14 +2047,14 @@ exportSnapshot sesh snap destinationPath = do
         let sourcePath = Paths.getNamedSnapshotDir snapDir
 
         destinationExists <- FS.doesDirectoryExist hfs destinationPath
-        when destinationExists $ throwIO (SnapshotExportDirExistsError sourcePath)
+        when destinationExists $ throwIO $ ErrSnapshotExportDirExists $ snapshotFsErrorPath hfs mode destinationPath
 
         withRollback_ reg
           (FS.createDirectoryIfMissing hfs True destinationPath)
           (FS.removeDirectoryRecursive hfs destinationPath)
 
-        -- Create hard links for all files in the destination directory
-        FS.hardLinkDirectoryRecursive hfs hbio reg sourcePath destinationPath
+        -- export the files for the snapshot, either by hard linking or copying
+        FS.hardLinkOrCopyDirectoryRecursive hfs (fsMode hbio mode) reg sourcePath destinationPath
 
         -- Make the directory and its contents durable.
         FS.synchroniseDirectoryRecursive hfs hbio destinationPath